replace get-onboarded-context-by-id with get-onboarded-context-by-context-id
